+++ R/correlation.R
@@ -551,7 +551,15 @@ correlation <- function(data,
     include_factors <- TRUE
   }
 
-  if (method == "polychoric") multilevel <- TRUE
+  # definitely need factors for polychoric
+  if (method == "polychoric") {
+    multilevel <- TRUE
+    # convert all input to factors, but only if all input currently is numeric
+    # we allow mix of numeric and factors
+    if (all(vapply(data, is.numeric, FUN.VALUE = TRUE))) {
+      data <- datawizard::to_factor(data)
+    }
+  }
 
   # Clean data and get combinations -------------